refactor: split query schema resolver components
